+++ backend/src/resolver/pageResolver.js
@@ -1,4 +1,5 @@
 import pageService from '../service/pageService';
+import authService from "../service/authService";
 
 export default {
   Query: {
@@ -6,14 +7,20 @@ export default {
     page: (parent, args) => {return pageService.getPage(args._id)}
   },
   Mutation: {
-    createPage: (parent, args) => {
-      return pageService.createPage(args.page);
+    createPage: (parent, args, context) => {
+      if (authService.verifyToken(context.req, context.res, context.next)) {
+        return pageService.createPage(args.page);
+      }
     },
-    updatePage: (parent, args) => {
-      return pageService.setPage(args.page);
+    updatePage: (parent, args, context) => {
+      if (authService.verifyToken(context.req, context.res, context.next)) {
+        return pageService.setPage(args.page);
+      }
     },
-    deletePage: (parent, args) => {
-      return pageService.deletePage(args._id).then(result => result.deletedCount > 0);
+    deletePage: (parent, args, context) => {
+      if (authService.verifyToken(context.req, context.res, context.next)) {
+        return pageService.deletePage(args._id).then(result => result.deletedCount > 0);
+      }
     }
   }
 }
